requirements.txt
alembic==1.4.3
attrs==20.3.0
Babel==2.9.0
bcrypt==3.2.0
blinker==1.4
Cerberus==1.3.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
ecdsa==0.16.1
email-validator==1.1.2
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Principal==0.4.0
Flask-Security==3.0.0
flask-serialize==1.4.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
passlib==1.7.4
permissive-dict==1.0.3
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.0
pyotp==2.4.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
simplejson==3.17.2
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.22
twilio==6.50.1
urllib3==1.26.2
validator-collection==1.5.0
Werkzeug==1.0.1
WTForms==2.3.3
